Refactor the flash firmware dialog.
