+++ src/components/catalog/keyboard/firmware/CatalogFirmware.tsx
@@ -18,7 +18,6 @@ import {
 import moment from 'moment';
 import { sendEventToGoogleAnalytics } from '../../../../utils/GoogleAnalytics';
 import { ICatalogPhase } from '../../../../store/state';
-import FlashFirmwareDialog from './FlashFirmwareDialog.container';
 import { hexadecimal } from '../../../../utils/StringUtils';
 
 type CatalogFirmwareState = {
@@ -48,10 +47,6 @@ export default class CatalogFirmware extends React.Component<
       });
       return;
     }
-  }
-
-  onCloseFlashFirmwareDialog() {
-    this.props.flashFirmwareDialog!.close();
   }
 
   onClickFlash(firmware: IFirmware) {
@@ -123,9 +118,6 @@ export default class CatalogFirmware extends React.Component<
             )}
           </div>
         </div>
-        <FlashFirmwareDialog
-          onClose={this.onCloseFlashFirmwareDialog.bind(this)}
-        />
       </React.Fragment>
     );
   }
