+++ src/components/common/firmware/FlashFirmwareDialog.scss
@@ -1,4 +1,4 @@
-@import '../../../../variables';
+@import '../../../variables';
 
 .flash-firmware-dialog {
   &-info {
